Complex layers

+++ src/settings.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <stdlib.h> 
 //typedef double real_t;
-typedef int int_t;
+typedef size_t int_t;
